fix: `upgradeInstanceManager`
