+++ internal/controller/cluster_upgrade.go
@@ -21,9 +21,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	neturl "net/url"
 	"reflect"
 
 	corev1 "k8s.io/api/core/v1"
@@ -33,10 +30,8 @@ import (
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/configuration"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/url"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/postgres"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/persistentvolumeclaim"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/resources/instance"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
@@ -743,7 +738,7 @@ func (r *ClusterReconciler) upgradeInstanceManager(
 				}
 			}
 
-			err = upgradeInstanceManagerOnPod(ctx, r.StatusClient, postgresqlStatus.Pod, targetManager)
+			err = r.StatusClient.UpgradeInstanceManager(ctx, postgresqlStatus.Pod, targetManager)
 			if err != nil {
 				enrichedError := fmt.Errorf("while upgrading instance manager on %s (hash: %s): %w",
 					postgresqlStatus.Pod.Name,
@@ -766,57 +761,4 @@ func (r *ClusterReconciler) upgradeInstanceManager(
 	}
 
 	return nil
-}
-
-// upgradeInstanceManagerOnPod upgrades an instance manager of a Pod via an HTTP PUT request.
-func upgradeInstanceManagerOnPod(
-	ctx context.Context,
-	statusClient *instance.StatusClient,
-	pod *corev1.Pod,
-	targetManager *utils.AvailableArchitecture,
-) error {
-	binaryFileStream, err := targetManager.FileStream()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		err = binaryFileStream.Close()
-	}()
-
-	scheme := instance.GetStatusSchemeFromPod(pod)
-	updateURL := url.Build(scheme, pod.Status.PodIP, url.PathUpdate, url.StatusPort)
-	req, err := http.NewRequestWithContext(ctx, http.MethodPut, updateURL, nil)
-	if err != nil {
-		return err
-	}
-	req.Body = binaryFileStream
-
-	resp, err := statusClient.Client.Do(req)
-	if err != nil {
-		if errors.Is(err.(*neturl.Error).Err, io.EOF) {
-			// This is perfectly fine as the instance manager will
-			// synchronously update and this call won't return.
-			return nil
-		}
-
-		return err
-	}
-
-	if resp.StatusCode == http.StatusOK {
-		// This should not happen. See previous block.
-		return nil
-	}
-
-	var body []byte
-	body, err = io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	err = resp.Body.Close()
-	if err != nil {
-		return err
-	}
-
-	return fmt.Errorf(string(body))
 }
